feat: create route structure and placeholder for category list page
Modularized existing code from the category list page, improving organization and reusability.
