+++ src/main.jsx
@@ -3,12 +3,12 @@ import { createRoot } from "react-dom/client";
 import { ThemeProvider } from "@mui/material";
 import theme from "./styles/theme.js";
 import App from "./App.jsx";
-import './styles/index.css'
+import './styles/index.css';
 
 createRoot(document.getElementById("root")).render(
   <React.StrictMode>
     <ThemeProvider theme={theme}> 
       <App />
     </ThemeProvider>
-  </React.StrictMode>,
+  </React.StrictMode>
 );
